ai-prompt-firewall/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
py-modules = ["config", "models"]

[tool.setuptools.packages.find]
include = ["api*", "audit_logging*", "config*", "dashboard*", "data*", "detection*", "engine*"]

[tool.setuptools.package-data]
engine = ["rules/*.yaml"]

[project]
name = "ai-prompt-firewall"
version = "1.0.0"
description = "AI Prompt Firewall — intercepts, analyzes, and blocks adversarial LLM prompts"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "sentence-transformers>=2.6",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx",  # for TestClient
    "ruff>=0.3",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
